requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0        # required by FastAPI TestClient for async transports
